requirements.txt
# Core app deps
streamlit
pandas
python-dotenv

# LLM / extraction
langchain-groq
langchain-core
langchain-ollama
langchain-community
pydantic

# Agentic AI (Agno framework)
agno
duckduckgo-search
yfinance
google-generativeai

# Vector Database
chromadb
sentence-transformers

# MCP (Model Context Protocol)
mcp

# Evaluation & Testing
rouge-score
nltk

# Ethics & Safety
fairlearn
presidio-analyzer
presidio-anonymizer
codecarbon

# Utilities
requests
